+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * Python: Added BITPOS command ([#1604](https://github.com/aws/glide-for-redis/pull/1604))
 * Python: Added GETEX command ([#1612](https://github.com/aws/glide-for-redis/pull/1612))
 * Python: Added BITFIELD and BITFIELD_RO commands ([#1615](https://github.com/aws/glide-for-redis/pull/1615))
+* Python: Added ZREVRANK command ([#1614](https://github.com/aws/glide-for-redis/pull/1614))
 
 ### Breaking Changes
 * Node: Update XREAD to return a Map of Map ([#1494](https://github.com/aws/glide-for-redis/pull/1494))